refactor(frontend): complete features/podcasts/ui module (FR-006 US2.4)
