+++ frontend/src/app/(protected)/podcasts/page.tsx
@@ -1,4 +1,4 @@
-import { PodcastList } from '@/components/podcasts/PodcastList';
+import { PodcastList } from '@/features/podcasts/ui/PodcastList';
 
 export default function PodcastsPage() {
   return <PodcastList />;
